refactor(LC030): split singleton fixer helpers
- separate field/property rewrites, member usage rewriting, and syntax helpers
- keep fixer behavior unchanged while reducing local complexity
